[td-198] super table interval query join without ts matched.

+++ src/client/inc/tsclient.h
@@ -196,7 +196,7 @@ typedef struct SDataBlockList {
 
 typedef struct SQueryInfo {
   int16_t  command;  // the command may be different for each subclause, so keep it seperately.
-  uint16_t type;     // query/insert/import type
+  uint32_t type;     // query/insert/import type
   char     slidingTimeUnit;
 
   int64_t         etime, stime;
